requirements.txt
Django==4.1
djangorestframework==3.14
psycopg2==2.9.3
celery==5.2.7
joblib==1.1.0
numpy==1.21.2
scikit-learn==0.24.2
tensorflow==2.7.0
redis==4.0.2
python-dotenv==0.19.0
